feat: mock tags
and poster -> request.ts
