--- lib/poster.ts
@@ -1,9 +1,0 @@
-export const poster = (route: string, body: any) =>
-  fetch('/api' + route, {
-    credentials: 'include',
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify(body),
-  });
