+++ mock/controllers/controllerMocks.js
@@ -9,6 +9,7 @@ const stubResponse = serviceMocks.stubResponse;
 router.get('/profile', (req, res) => mapMockResponse(req, res));
 router.get('/posts', (req, res) => mapMockResponse(req, res));
 router.get('/post/*', (req, res) => mapMockResponse(req, res));
+router.get('/tags', (req, res) => mapMockResponse(req, res));
 
 router.post('/login', (req, res) => stubResponse(req, res));
 router.post('/register', (req, res) => stubResponse(req, res));
