+++ package.json
@@ -1,58 +1,59 @@
 {
-  "name": "devfit-nextjs",
-  "version": "0.0.2",
-  "license": "MIT",
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "tc-dev": "nodemon server/index.js",
-    "tc-start": "cross-env NODE_ENV=production node server/index.js",
-    "db": "prisma studio --experimental",
-    "db-save": "prisma migrate save --experimental",
-    "db-up": "prisma migrate up --experimental",
-    "update": "prisma generate",
-    "go": "npm run db-save && npm run db-up && npm run update",
-    "postinstall": "npm run update"
-  },
-  "dependencies": {
-    "@emotion/core": "~10.0.28",
-    "@kevinwolf/formal-web": "~0.1.0",
-    "@prisma/client": "~2.0.0-beta.2",
-    "@zeit/next-css": "~1.0.1",
-    "apollo-boost": "~0.4.7",
-    "apollo-client": "~2.6.8",
-    "bcryptjs": "~2.4.3",
-    "classnames": "~2.2.6",
-    "cookie": "~0.4.0",
-    "date-fns": "~1.30.1",
-    "express": "~4.16.4",
-    "file-loader": "~3.0.1",
-    "graphql": "~14.6.0",
-    "http-proxy-middleware": "~0.19.1",
-    "isomorphic-unfetch": "~3.0.0",
-    "next": "~9.3.3",
-    "react": "~16.13.0",
-    "react-apollo-hooks": "~0.5.0",
-    "react-dom": "~16.13.0",
-    "react-infinite-scroller": "~1.2.4",
-    "react-md-spinner": "~0.3.0",
-    "react-modal": "~3.8.1",
-    "semantic-ui-react": "~0.88.2",
-    "webpack": "4.4.0",
-    "yup": "~0.28.3"
-  },
-  "devDependencies": {
-    "@prisma/cli": "~2.0.0-beta.2",
-    "@types/bcryptjs": "~2.4.2",
-    "@types/cookie": "~0.3.3",
-    "@types/yup": "~0.26.35",
-    "@emotion/core": "~10.0.28",
-    "cross-env": "~5.2.0",
-    "nodemon": "~1.18.11",
-    "@types/node": "~13.11.1",
-    "typescript": "~3.8.3",
-    "semantic-ui-react": "~0.88.2",
-    "@types/react": "~16.9.34"
-  }
+    "name": "devfit-nextjs",
+    "version": "0.0.2",
+    "license": "MIT",
+    "scripts": {
+        "dev": "next dev",
+        "build": "next build",
+        "start": "next start",
+        "start-prod": "cross-env NODE_ENV=production next start",
+        "tc-dev": "nodemon server/index.js",
+        "tc-start": "cross-env NODE_ENV=production node server/index.js",
+        "db": "prisma studio --experimental",
+        "db-save": "prisma migrate save --experimental",
+        "db-up": "prisma migrate up --experimental",
+        "update": "prisma generate",
+        "go": "npm run db-save && npm run db-up && npm run update",
+        "postinstall": "npm run update"
+    },
+    "dependencies": {
+        "@emotion/core": "~10.0.28",
+        "@kevinwolf/formal-web": "~0.1.0",
+        "@prisma/client": "~2.0.0-beta.2",
+        "@zeit/next-css": "~1.0.1",
+        "apollo-boost": "~0.4.7",
+        "apollo-client": "~2.6.8",
+        "bcryptjs": "~2.4.3",
+        "classnames": "~2.2.6",
+        "cookie": "~0.4.0",
+        "date-fns": "~1.30.1",
+        "express": "~4.16.4",
+        "file-loader": "~3.0.1",
+        "graphql": "~14.6.0",
+        "http-proxy-middleware": "~0.19.1",
+        "isomorphic-unfetch": "~3.0.0",
+        "next": "~9.3.3",
+        "react": "~16.13.0",
+        "react-apollo-hooks": "~0.5.0",
+        "react-dom": "~16.13.0",
+        "react-infinite-scroller": "~1.2.4",
+        "react-md-spinner": "~0.3.0",
+        "react-modal": "~3.8.1",
+        "semantic-ui-react": "~0.88.2",
+        "webpack": "4.4.0",
+        "yup": "~0.28.3"
+    },
+    "devDependencies": {
+        "@prisma/cli": "~2.0.0-beta.2",
+        "@types/bcryptjs": "~2.4.2",
+        "@types/cookie": "~0.3.3",
+        "@types/yup": "~0.26.35",
+        "@emotion/core": "~10.0.28",
+        "cross-env": "~5.2.0",
+        "nodemon": "~1.18.11",
+        "@types/node": "~13.11.1",
+        "typescript": "~3.8.3",
+        "semantic-ui-react": "~0.88.2",
+        "@types/react": "~16.9.34"
+    }
 }
